+++ PSTibber.Tests.ps1
@@ -7,22 +7,22 @@ $global:Pester_Address = 'Winterfell Castle 1'
 $global:Pester_EAN = '735999102107573183'
 $global:Pester_Email = 'arya@winterfell.com'
 
-Describe "Invoke-TibberGraphQLQuery" {
+Describe "Invoke-TibberQuery" {
     It "Can invoke GraphQL query" {
         $query = "{ viewer{ homes{ id } } }"
-        (Invoke-TibberGraphQLQuery -Query $query).viewer.homes[0].id | Should -Be $Pester_HomeId
+        (Invoke-TibberQuery -Query $query).viewer.homes[0].id | Should -Be $Pester_HomeId
     }
 
     It "Fails when invalid query" {
-        { Invoke-TibberGraphQLQuery -Query "{}" } | Should -Throw
+        { Invoke-TibberQuery -Query "{}" } | Should -Throw
     }
 
     It "Fails when invalid query data" {
-        Invoke-TibberGraphQLQuery -Query "{ viewer{ home(id:`"$Pester_HomeId_`"){ id }}}" -ErrorAction Ignore | Should -Be $null
+        Invoke-TibberQuery -Query "{ viewer{ home(id:`"$Pester_HomeId_`"){ id }}}" -ErrorAction Ignore | Should -Be $null
     }
 
     It "Fails when invalid URI" {
-        { Invoke-TibberGraphQLQuery -URI 'https://api.tibber.com/v1-alpha/gql' -Query "{}" } | Should -Throw
+        { Invoke-TibberQuery -URI 'https://api.tibber.com/v1-alpha/gql' -Query "{}" } | Should -Throw
     }
 }
 
@@ -93,5 +93,37 @@ Describe "Get-TibberPriceInfo" {
 
     It "Can get power production (w/ access token)" {
         Get-TibberPriceInfo -PersonalAccessToken $Pester_AccessToken -HomeId $Pester_HomeId -Last 100 | Should -Not -Be $null
+    }
+}
+
+Describe "Connect-TibberWebSocket" {
+    It "Can connect WebSocket" {
+        $global:connection = Connect-TibberWebSocket
+        $connection.WebSocket | Should -Not -Be $null
+    }
+}
+
+Describe "Register-TibberLiveConsumptionSubscription" {
+    It "Can register subscription" {
+        $global:subscription = Register-TibberLiveConsumptionSubscription -Connection $connection -HomeId $Pester_HomeId
+        $subscription.Id | Should -Not -Be $null
+    }
+}
+
+Describe "Read-TibberWebSocket" {
+    It "Can read from WebSocket" {
+        Read-TibberWebSocket -Connection $connection -Callback {} -Count 1 | Should -Not -Be $null
+    }
+}
+
+Describe "Unregister-TibberLiveConsumptionSubscription" {
+    It "Can unregister subscription" {
+        { Unregister-TibberLiveConsumptionSubscription -Connection $connection -Subscription $subscription } | Should -Not -Throw
+    }
+}
+
+Describe "Disconnect-TibberWebSocket" {
+    It "Can disconnect WebSocket" {
+        { Disconnect-TibberWebSocket -Connection $connection } | Should -Not -Throw
     }
 }